+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [BREAKING] Replaced `.enqueue` (only supported sidekiq) with `.call_async` (via a configurable registry of backgrounding libraries)
 * [FEAT] attachable now creates a foo_async to call call_async
 * `type` validator is not still applied to the blank value when allow_blank is true (`type: Hash` will no longer accept `false` or `""`)
+* [FEAT] `expects`/`exposes` now prefers new `optional: true` over allow_blank for simplicity
 
 ## 0.1.0-alpha.2.8.1
 * [BUGFIX] Fixed symbol callback and message handlers not working in inherited classes due to private method visibility issues